Something like this:

```
    benchmark                old ns/op     new ns/op     delta
    BenchmarkIndexLoad-4     517378116     81028817      -84.34%
    BenchmarkOpen-4          519653111     79758368      -84.65%
    BenchmarkWrite-4         125716        121013        -3.74%
    BenchmarkRead-4          2349          2118          -9.83%
```